chore: move fn to util
